+++ packages/svelte-playwright-coverage/src/spc/convert/index.ts
