+++ packages/svelte-playwright-coverage/src/spc/preprocess.ts
@@ -1,4 +1,5 @@
 import kleur from 'kleur'
+import path from 'node:path'
 import { APP_NAME } from '../constants.js'
 import { OutDir } from '../utils/OutDir.js'
 import { TmpDir } from '../utils/TmpDir.js'
@@ -17,4 +18,10 @@ ${kleur.bold(`☂️ ${APP_NAME}`)}
   await Promise.all([TmpDir.generate(), OutDir.set(output)])
 
   console.log(kleur.cyan('Measuring coverage...'))
+
+  return {
+    root:
+      // process.cwd()
+      path.join(process.cwd(), 'packages/demo')
+  }
 }
