+++ packages/svelte-playwright-coverage/src/spc/types/Context.ts
@@ -1,0 +1,3 @@
+import { preprocess } from '../preprocess.js'
+
+export type Context = Awaited<ReturnType<typeof preprocess>>
