+++ packages/svelte-playwright-coverage/src/spc/types/ResolvedCoverage.ts
