+++ packages/svelte-playwright-coverage/src/spc/utils/thinning.ts
