+++ packages/svelte-playwright-coverage/src/spc/utils/transformFile.ts
